+++ NEWS.md
@@ -1,5 +1,7 @@
 # sharepointr (development version)
 
+* Add `upload_sp_items()` function (2024-06-24).
+
 # sharepointr 0.1.0
 
 * Initial version.